Refactor so system depedencies are installed one time at start of wbi
