+++ cmd/setup.go
@@ -32,8 +32,12 @@ func newSetup(setupOpts setupOpts) error {
 
 	var WBConfig config.WBConfig
 
-	// Determine OS
+	// Determine OS and install pre-requisites
 	osType, err := os.DetectOS()
+	if err != nil {
+		return err
+	}
+	err = os.InstallPrereqs(osType)
 	if err != nil {
 		return err
 	}
